код подведен под требования
